+++ cfgov/v1/models/blog_page.py
@@ -13,11 +13,13 @@ from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
 
 from . import AbstractFilterPage, CFGOVPageManager
 from ..atomic_elements import molecules, organisms
+from .. import blocks as v1_blocks
 
 class BlogPage(AbstractFilterPage):
     content = StreamField([
         ('full_width_text', organisms.FullWidthText()),
         ('image_text_50_50_group', organisms.ImageText5050Group()),
+        ('feedback', v1_blocks.Feedback()),
     ])
     edit_handler = AbstractFilterPage.generate_edit_handler(
         content_panel = StreamFieldPanel('content')
@@ -28,6 +30,7 @@ class BlogPage(AbstractFilterPage):
 class LegacyBlogPage(AbstractFilterPage):
     content = StreamField([
         ('content', blocks.RawHTMLBlock(help_text='Content from WordPress unescaped.')),
+        ('feedback', v1_blocks.Feedback()),
     ])
     objects = CFGOVPageManager()
     edit_handler = AbstractFilterPage.generate_edit_handler(